+++ src/main/java/com/dawid/ems/payload/UserSummary.java
@@ -2,19 +2,19 @@ package com.dawid.ems.payload;
 
 
 public class UserSummary {
-    private Long id;
+    private int id;
     private String username;
 
-    public UserSummary(Long id, String username) {
+    public UserSummary(int id, String username) {
         this.id = id;
         this.username = username;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
